tests for `help`, `log`, `clean`
